% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-related.R
\name{create_pkgcheck_config}
\alias{create_pkgcheck_config}
\title{create a pkgcheck config file}
\usage{
create_pkgcheck_config(libpaths = NULL, output = "output",
  loglevel = "info", .start_config_file = NULL, .overwrite = TRUE)
}
\arguments{
\item{libpaths}{vector of library paths}

\item{output}{output directory}

\item{loglevel}{log level}

\item{.start_config_file}{existing yaml configuration file}

\item{.overwrite}{whether to overwrite existing settings or just add new}
}
\description{
create a pkgcheck config file
}
\details{
log levels: debug, info, warn, fatal

if overwrite is false, any settings present in the pkgcheck.yml file
will be retained, and the parameter value will be ignored.
}
