% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_packrat_lockfile}
\alias{gen_packrat_lockfile}
\title{Wrapper function to help generate a packrat lockfile}
\usage{
gen_packrat_lockfile(..., .workdir = file.path(getwd(), "work"))
}
\arguments{
\item{...}{the arguments to pass to gen_runtime_description}

\item{.workdir}{directory to pass to install_from_desc}
}
\description{
Wrapper function to help generate a packrat lockfile
}
