% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_pkg_desc_from_libdir}
\alias{gen_pkg_desc_from_libdir}
\title{Build a string suitable for pasting into source code for passing to gen_runtime_description
    This is meant to help kickstart a new source file that will be checked into source control
    so that you can have a starter list.
    
    Caveate: this will essentially make all installed packages top-level citizens when calling 
    install.packages(dependencies=c("Depends", "Imports", "LinkingTo", "Suggests"))}
\usage{
gen_pkg_desc_from_libdir(.libdir, .addversion = FALSE)
}
\arguments{
\item{.libdir}{the library directory to scan}
}
\description{
Build a string suitable for pasting into source code for passing to gen_runtime_description
    This is meant to help kickstart a new source file that will be checked into source control
    so that you can have a starter list.
    
    Caveate: this will essentially make all installed packages top-level citizens when calling 
    install.packages(dependencies=c("Depends", "Imports", "LinkingTo", "Suggests"))
}
