% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_description.R
\name{gen_runtime_description}
\alias{gen_runtime_description}
\title{create decription file for internal package}
\usage{
gen_runtime_description(pkgs = NULL, github = NULL, bioc = NULL,
  git = NULL, url = NULL, name = NULL, email = NULL,
  runtime = paste0("runtime", gsub("-", "", as.Date.character(Sys.Date()))))
}
\arguments{
\item{pkgs}{packages}

\item{github}{github packages}

\item{bioc}{bioconductor packages}

\item{git}{git packages}

\item{url}{url packages}

\item{name}{name}

\item{email}{author email}

\item{runtime}{runtime name}
}
\value{
desc file
}
\description{
create decription file for internal package
}
\details{
creates the desc object representation with the fields
needed to bootstrap an internal package description

bootstrapping adds devtools and packrat to imports. If refined version(s) or
other tweaks needed set bootstrap to false, and directly set.
}
