% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_snapshot.R
\name{gen_snapshot}
\alias{gen_snapshot}
\title{generate a lockfile from a requested set of packages}
\usage{
gen_snapshot(pkgs, .dir = tempdir(), snapshot_tars = FALSE)
}
\arguments{
\item{pkgs}{packages to install}

\item{snapshot_tars}{snapshot the tars}
}
\description{
generate a lockfile from a requested set of packages
}
\details{
generate a packrat snapshot from a requested set of
packages by mocking a project folder and letting
packrat scan and generate a lockfile
}
