% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packrat_libpaths.R
\name{get_packrat_libpaths}
\alias{get_packrat_libpaths}
\title{get the libpaths associated with a project that uses packrat}
\usage{
get_packrat_libpaths(.pdir)
}
\arguments{
\item{.pdir}{packrat directory}
}
\description{
get the libpaths associated with a project that uses packrat
}
