% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_runtime.R
\name{install_from_desc}
\alias{install_from_desc}
\title{install from a runtime description}
\usage{
install_from_desc(.d, .dir = fs::path_temp())
}
\arguments{
\item{.d}{description file}

\item{.dir}{directory to install pkgs from description}
}
\description{
install from a runtime description
}
