% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_runtime.R
\name{install_runtime}
\alias{install_runtime}
\title{install a runtime from a given packrat lockfile}
\usage{
install_runtime(.rt, .libs = NULL)
}
\arguments{
\item{.rt}{location of a runtime template}

\item{.libs}{library locations to use when installing the runtime in isolation}
}
\value{
path to installed runtime packages
}
\description{
install a runtime from a given packrat lockfile
}
