% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_packrat_opts.R
\name{packrat_settings}
\alias{packrat_settings}
\title{packrat settings}
\usage{
packrat_settings(use.cache = TRUE, external.packages = NULL,
  local.repos = NULL, ignored.packages = "internal",
  snapshot.fields = c("Imports", "Depends", "LinkingTo", "Suggests"))
}
\arguments{
\item{use.cache}{use packrat cache when installing packages}

\item{external.packages}{packages to consider external to packrat}

\item{local.repos}{local cranlike repos to install from}

\item{ignored.packages}{packages to not snapshot}

\item{snapshot.fields}{which fields to capture in lockfile}
}
\description{
packrat settings
}
