% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_runtime.R
\name{setup_runtime_template}
\alias{setup_runtime_template}
\title{setup a runtime template}
\usage{
setup_runtime_template(.ll, .dir = fs::path_temp(), ...)
}
\arguments{
\item{.ll}{lockfile lines}

\item{.dir}{dir to create the runtime template}

\item{...}{parameters to pass to `packrat_settings()`}
}
\description{
setup a runtime template
}
