% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgcheck-related.R
\name{write_config}
\alias{write_config}
\title{write out a yaml config from a configuration list}
\usage{
write_config(.config, .file)
}
\arguments{
\item{.config}{configuration}

\item{.file}{yaml file, should be named pkgcheck.yaml}
}
\description{
write out a yaml config from a configuration list
}
